split usePokemon as hook folder
