+++ src/components/NameInput/NameInput.js
@@ -1,23 +1,8 @@
 import "./NameInput.scss"
-import { registerHtml, useState, useGlobalState, useEffect } from "tram-one"
-import Pokedex from 'pokedex-promise-v2'
-const PokeAPI = new Pokedex()
+import { registerHtml } from "tram-one"
+import usePokemon from '../../hooks/usePokemon'
 
 const html = registerHtml()
-
-const usePokemon = () => {
-  const [pokemonName, setPokemonName] = useState('')
-  const [pokemon, setPokemon] = useGlobalState('pokemon', null)
-
-  const onUpdatePokemonName = (event) => setPokemonName(event.target.value)
-
-  useEffect(async () => {
-    const fetchedPokemon = await PokeAPI.getPokemonByName(pokemonName.toLowerCase())
-    setPokemon(fetchedPokemon)
-  }, [pokemonName])
-
-  return { pokemon, pokemonName, onUpdatePokemonName }
-}
 
 export default () => {
   const {pokemonName, onUpdatePokemonName} = usePokemon()
