+++ src/components/PokeImage/PokeImage.js
@@ -1,13 +1,14 @@
 import "./PokeImage.scss"
-import { registerHtml, useGlobalState } from "tram-one"
+import { registerHtml } from "tram-one"
 import PokeBallImage from "./PokeBallImage"
+import usePokemon from '../../hooks/usePokemon'
 
 const html = registerHtml({
   PokeBallImage
 })
 
 export default () => {
-  const [pokemon] = useGlobalState('pokemon', null)
+  const { pokemon } = usePokemon()
 
   if (!pokemon) {
     return html`<PokeBallImage />`
