+++ src/components/PokemonTypes/PokemonTypes.js
@@ -1,12 +1,13 @@
 import { registerHtml, useGlobalState } from "tram-one"
 import TypeBadge from "../TypeBadge";
+import usePokemon from "../../hooks/usePokemon";
 
 const html = registerHtml({
   TypeBadge
 })
 
 export default () => {
-  const [pokemon] = useGlobalState('pokemon', null)
+  const { pokemon } = usePokemon()
 
   const types = pokemon && pokemon.types.map(({type: {name}}) => html`<TypeBadge type=${name}/>`)
